Rename package so the last portion of the package name is more specific+appropriate (so the lollipop path isn't /system/priv-app/roothelper/roothelper.apk) And remove spaces to prevent issues when compiling inline with AOSP
