+++ GSamBatteryMonitor-RootCompanion/.classpath
