+++ GSamBatteryMonitor-RootCompanion/.gitignore
