+++ GSamBatteryMonitor-RootCompanion/.project
