+++ GSamBatteryMonitor-RootCompanion/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.gsamlabs.bbm.rootcompanion"
+    package="com.gsamlabs.bbm.GSamBatteryMonitorRootHelper"
     android:versionCode="4"
     android:versionName="1.3" >
 
@@ -31,7 +31,7 @@
         android:theme="@style/AppBaseTheme" >
         <!-- NOTE:  This activity may get disabled, thereby hiding the app from the launcher -->
         <activity
-            android:name="com.gsamlabs.bbm.rootcompanion.MainActivityLauncher"
+            android:name="com.gsamlabs.bbm.GSamBatteryMonitorRootHelper.MainActivityLauncher"
             android:label="@string/app_name" 
             android:noHistory="true">
             <intent-filter>
@@ -41,7 +41,7 @@
         </activity>
         
         <activity
-            android:name="com.gsamlabs.bbm.rootcompanion.MainActivity"
+            android:name="com.gsamlabs.bbm.GSamBatteryMonitorRootHelper.MainActivity"
             android:label="@string/app_name"
             android:exported="true" >
             <intent-filter>
@@ -52,10 +52,10 @@
 
         <!-- The service providing getStatistics - we require our custom BATTERY_STATS permission -->
         <service
-            android:name="com.gsamlabs.bbm.rootcompanion.BatteryStatsService"
+            android:name="com.gsamlabs.bbm.GSamBatteryMonitorRootHelper.BatteryStatsService"
             android:permission="com.gsamlabs.permission.BATTERY_STATS" >
             <intent-filter>
-                <action android:name="com.gsamlabs.bbm.rootcompanion.IBatteryStatsService" />
+                <action android:name="com.gsamlabs.bbm.GSamBatteryMonitorRootHelper.IBatteryStatsService" />
             </intent-filter>
         </service>
     </application>
