+++ GSamBatteryMonitor-RootCompanion/proguard-project.txt
