+++ GSamBatteryMonitor-RootCompanion/project.properties
