+++ GSamBatteryMonitor-RootCompanion/res/drawable-hdpi/ic_launcher.png
