+++ GSamBatteryMonitor-RootCompanion/res/drawable-ldpi/ic_launcher.png
