+++ GSamBatteryMonitor-RootCompanion/res/drawable-mdpi/ic_launcher.png
