+++ GSamBatteryMonitor-RootCompanion/res/drawable-xhdpi/ic_launcher.png
