+++ GSamBatteryMonitor-RootCompanion/res/drawable-xxhdpi/ic_launcher.png
