+++ GSamBatteryMonitor-RootCompanion/res/layout/activity_main.xml
