+++ GSamBatteryMonitor-RootCompanion/res/values/strings.xml
