+++ GSamBatteryMonitor-RootCompanion/res/values/styles.xml
