+++ GSamBatteryMonitor-RootCompanion/src/com/gsamlabs/bbm/GSamBatteryMonitorRootHelper/IBatteryStatsService.aidl
@@ -1,4 +1,4 @@
-package com.gsamlabs.bbm.rootcompanion;
+package com.gsamlabs.bbm.GSamBatteryMonitorRootHelper;
 
 interface IBatteryStatsService
 {
